Time: 69 ms (89.95%), Space: 79.5 MB (89.08%) - LeetHub

+++ 0452-minimum-number-of-arrows-to-burst-balloons/0452-minimum-number-of-arrows-to-burst-balloons.java
@@ -2,12 +2,12 @@ class Solution {
     public int findMinArrowShots(int[][] points) {
         
         Arrays.sort(points, (a1, a2)-> Integer.compare(a1[0],a2[0])); //[[1,6],[2,8],[7,12],[10,16]]
-        long currentEnd = Long.MAX_VALUE;
+        int currentEnd = Integer.MAX_VALUE;
         
         int minArrows = 1;
         for(int[] point: points){
-            long pointStart = point[0];
-            long pointEnd = point[1];
+            int pointStart = point[0];
+            int pointEnd = point[1];
             
             if(currentEnd >= pointStart){
                 // System.out.print(point[0]+" "+point[1]+" ");
